# openapi/swagger.yaml
openapi: 3.0.0
info:
  title: WeHOME
  version: 1.0.0
  description: Solve Accommodation Problems
  termsOfService: http://swagger.io/terms/
  contact:
    email: 19020024@vnu.edu.vn
...
servers:
  - url: http://wehome-env.eba-ujwahtza.ap-southeast-1.elasticbeanstalk.com/
  - url: http://localhost:8889
  - url: http://113.167.196.55:8889
...
paths:
...
  /api/room:
    get:
      tags:
        - room
      summary: Get all room
      operationId: room.getRooms
      parameters:
        - name: limit
          in: query
          description: number of items per page
          required: false
          schema:
            type: number
        - name: page
          in: query
          description: index of page
          required: false
          schema:
            type: number
      responses:
        '500':
          description: System error
        default:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /api/room/{roomId}:
    get:
      tags:
        - room
      summary: Get room by id
      operationId: room.getRoomById
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Invalid roomId supplied
          content: { }
        '404':
          description: Room not found
          content: { }
        '500':
          description: System error
    put:
      tags:
        - room
      security:
        - bearerAuth: [ ]
      summary: Update room
      operationId: room.updateRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        description: Updated room object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid roomId supplied
          content: { }
        '404':
          description: Room not found
          content: { }
        '500':
          description: System error
    delete:
      tags:
        - room
      security:
        - bearerAuth: [ ]
      summary: Delete room
      operationId: room.deleteRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: number
      responses:
        '401':
          description: Access token is missing or invalid
        '200':
          description: OK
          content: { }
        '400':
          description: Invalid roomId supplied
          content: { }
        '404':
          description: Room not found
          content: { }
        '500':
          description: System error
  /api/room/create:
    post:
      tags:
        - room
      security:
        - bearerAuth: [ ]
      summary: Create room
      operationId: room.createRoom
      requestBody:
        description: Created room object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '401':
          description: Access token is missing or invalid
        '500':
          description: System error
        default:
          description: OK
  /api/room/search:
    post:
      tags:
        - room
      summary: Search room by geo coordinate
      operationId: room.search
      parameters:
        - name: limit
          in: query
          description: number of items per page
          required: false
          schema:
            type: number
        - name: page
          in: query
          description: index of page
          required: false
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
                radius:
                  type: number
                  format: float
                  description: km
      responses:
        '500':
          description: System error
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /api/room/filter:
    post:
      tags:
        - room
      summary: Filter room by host ID
      operationId: room.filterRoom
      parameters:
        - name: limit
          in: query
          description: number of items per page
          required: false
          schema:
            type: number
        - name: page
          in: query
          description: index of page
          required: false
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hostId:
                  type: number
                filter:
                  type: string
                  enum:
                    - Empty
                    - Arriving soon
                    - Checking out
                    - Currently hosting
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: No room
          content: { }
        '500':
          description: System error
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Room:
      type: object
      properties:
        roomId:
          type: number
        name:
          type: string
        address_id:
          type: number
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        roomType:
          type: number
        hostId:
          type: number
        numGuest:
          type: integer
        numBed:
          type: integer
        numBedroom:
          type: integer
        numBathroom:
          type: integer
        rule:
          type: string
          nullable: true
        accommodationType:
          type: string
        price:
          type: number
        confirmed:
          type: boolean
          default: false
        rate:
          type: number
          format: float
          nullable: true
        image:
          type: array
          items:
            type: string
        facility:
          type: array
          items:
            type: string
      example:
        room_id: 102
        room_name: Family Room with balcony (Ruby Hotel Bac Ninh)
        latitude: 20.0313
        longitude: 105.863
        address_id: 6
        roomType: 2
        numGuest: 4
        numBed: 2
        numBedroom: 2
        numBathroom: 3
        rule: 'Nhận phòng: Sau 13:00;Trả phòng: 13:00;Không thú cưng;Cho phép hút thuốc'
        accommodationType: Toàn bộ nhà
        confirmed: true
        rate: 3.8
        host_id: 86
        price: 2400000
        image:
          - https://a0.muscache.com/im/pictures/47c9affe-c88d-438e-bbc2-63da61d51305.jpg?im_w=720
          - https://a0.muscache.com/im/pictures/98b73564-2d61-4ac8-b90d-966c07fe9695.jpg?im_w=720
        facility:
            - wifi
            - Chỗ để ô tô
            - Dịch vụ dọn phòng
...